Update tsconfig.json paths and add new TypeScript file
Updated the `outFile` paths in `tsconfig.json` to reflect a new
directory structure. Specifically, the paths for `Snebur.js` and
`Snebur.UI.js` have been changed to the new locations.

Added a new file `src/Comunicacao/Utilidade/AjaxUtil.HttpRequest.ts`
to the list of files in the `tsconfig.json`.

Updated the `files` array to reflect the new path for `Snebur.d.ts`.